+++ src/core/prelude/date/create/index.ts
@@ -77,13 +77,14 @@ extend(Date, 'create', (pattern?: DateCreateValue) => {
 		};
 
 		const replacer = (str, date, time, zone) => {
-			const
-				normalizedDate = normalizeDate(date);
+			const normalizedDate = normalizeDate(date);
 
 			time = Object.isTruly(time) ? time : '00:00:00';
-			zone = Object.isTruly(zone) ? zone : getZone(normalizedDate);
 
-			return `${normalizedDate}T${time}${normalizeZone(zone)}`;
+			const dateTime = `${normalizedDate}T${time}`;
+			zone = Object.isTruly(zone) ? zone : getZone(dateTime);
+
+			return `${dateTime}${normalizeZone(zone)}`;
 		};
 
 		return new Date(Date.parse(pattern.replace(isDateStr, replacer)));